refactor(mitm-addon): extract tcp logging owner
